refactor: ImageIcon + keyValueInput
