+++ src/components/atoms/ImageIcon/ImageIcon.styled.ts
@@ -1,6 +1,8 @@
 import styled from 'styled-components';
 
-export const StyledImageIcon = styled.img<{height?: number; width?: number}>`
-  height: ${props => (props.height ? `${props.height}px` : '32px')}};
-  width: ${props => (props.width ? `${props.width}px` : '32px')}};
-  `;
+export const StyledImageIcon = styled.img<{$height?: number; $width?: number}>`
+  ${({$height, $width}) => `
+    height: ${$height || '32'}px;
+    width: ${$width || '32'}px;
+`}
+`;
