+++ src/components/atoms/KeyValueInput/KeyValueInput.styled.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  max-height: 800px;
+  overflow-y: auto;
+`;
+
+export const TitleContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+export const TitleLabel = styled.span``;
